Moving messages validator into app namespace
