+++ app/validators/messages.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-i18n-cp-validations/validators/messages';
